saving rates done... fixed the issue when deleting dates from hours

+++ database/migrations/2025_10_07_234100_create_shifts_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->unsignedMediumInteger('user_id');
             $table->time('start');
             $table->time('end');
-            $table->integer('break')->nullable();
+            $table->decimal('break', 5, 2)->nullable();
 
             $table->unique(['date', 'user_id']);
